raspberry_pi/requirements.txt
fastapi
uvicorn[standard]
pyserial
rplidar-roboticia
picamera2
